requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.11.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
Pillow>=10.4.0
edge-tts>=6.1.10
gTTS>=2.5.3
yt-dlp>=2026.1.0
